Added a few plugins.
i added a few plugins and gave them some config files. the plugins that were installed and what they are specified in the comments of the plugins.lua file. descriptions for what they keybidings are and what they do should be in the comment next to where the key binding is defined.